pyproject.toml
[project]
name = "eyes-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.56.0",
    "groq>=1.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
]
